requirements.txt
# requirements.txt
streamlit
google-generativeai
python-dotenv
requests
gtts